add resolvers for users
