+++ src/resolvers/index.js
@@ -1,0 +1,7 @@
+const Query = require('./Query');
+const Mutation = require('./Mutation');
+
+module.exports = {
+  Query,
+  Mutation
+};
